my-finances

+++ src/app/api/deleteTransaction/route.tsx
@@ -10,6 +10,7 @@ export async function GET(request: Request) {
     const id = urlParams.get('id');     
     const from_transaction = urlParams.get('from_transaction');     
     const which = urlParams.get('which');     
+    const date = urlParams.get('date');     
 
     const data = await sql`
         SELECT id FROM users
@@ -19,15 +20,19 @@ export async function GET(request: Request) {
 
     let response;
 
-    if (data.rows.length && id) {
+    if (data.rows.length && id && date) {
 
         const execute = await sql`
             DELETE FROM transactions
             WHERE 
             (user_id = ${data.rows[0].id} AND id = ${id})
-            OR (from_transaction = ${from_transaction} AND ${which} = 'all')
+            OR (from_transaction = ${from_transaction} AND ${which} = 'all' AND date >= ${date})
         `;
 
+        if (which == 'all') {
+            await sql`UPDATE transactions SET recurring = false WHERE user_id = ${data.rows[0].id} AND from_transaction = ${from_transaction}`;
+        }
+        
         if (execute) {
             response = {
                 code: 1,
